Migrate Warehouse Exclusive Box States ("LOST" "SCRAP") to True States

+++ mobile/savebox.php
@@ -27,7 +27,22 @@ if (!$_POST['qr_id']) {
 }
 // Box creation/update
 [$new, $box, $message] = db_transaction(function () use ($new) {
-    $box = db_row('SELECT * FROM stock WHERE id = :id', ['id' => $_POST['id']]);
+    $box = db_row('SELECT 
+                stock.*, 
+                locations.type as locationType,
+                DATE_FORMAT(stock.modified,"%Y/%m/%d") AS statemodified,
+                bs.id as box_state_id, 
+                bs.label as box_state_name
+                FROM stock
+                INNER JOIN box_state bs ON bs.id = stock.box_state_id
+                LEFT OUTER JOIN locations ON stock.location_id = locations.id 
+                WHERE stock.id = :id', ['id' => $_POST['id']]);
+
+    $is_scrap = (!empty($_POST['scrap']));
+    $is_lost = (!empty($_POST['lost']));
+
+    // Getting the new box state id based on the location
+    $newboxstate = db_row('SELECT bs.id as box_state_id, bs.label as box_state_name FROM locations l INNER JOIN box_state bs ON bs.id = l.box_state_id WHERE l.id = :id', ['id' => $_POST['location_id']]);
 
     // Updates and Preparation
     if ($new) {
@@ -40,10 +55,31 @@ if (!$_POST['qr_id']) {
             }
         );
     } else {
-        // unset order state
-        if ($box['location_id'] != $_POST['location_id']) {
-            db_query('UPDATE stock SET ordered = NULL, ordered_by = NULL, picked = NULL, picked_by = NULL WHERE id = :id', ['id' => $box['id']]);
-            db_query('INSERT INTO itemsout (product_id, size_id, count, movedate, from_location, to_location) VALUES ('.$box['product_id'].','.$box['size_id'].','.$box['items'].',NOW(),'.$box['location_id'].','.$_POST['location_id'].')');
+        $id = $_POST['id'];
+
+        //  when checked scrap or lost in the form
+        if ($is_scrap) {
+            $newboxstate['box_state_id'] = 6;
+            $newboxstate['box_state_name'] = 'Scrap';
+        } elseif ($is_lost) {
+            $newboxstate['box_state_id'] = 2;
+            $newboxstate['box_state_name'] = 'Lost';
+        }
+
+        if ($box['location_id'] != $_POST['location_id'] && !$is_scrap && !$is_lost) {
+            // Boxes should not be relocated to virtual locations
+            // Related to https://trello.com/c/Ci74t1Wj
+            if (in_array($newboxstate['box_state_name'], ['Lost', 'Scrap'])) {
+                $_POST['location_id'] = $box['location_id'];
+                if ('Lost' == $newboxstate['box_state_name']) {
+                    $is_lost = true;
+                } elseif ('Scrap' == $newboxstate['box_state_name']) {
+                    $is_scrap = true;
+                }
+            } else {
+                db_query('UPDATE stock SET ordered = NULL, ordered_by = NULL, picked = NULL, picked_by = NULL WHERE id = :id', ['id' => $box['id']]);
+                db_query('INSERT INTO itemsout (product_id, size_id, count, movedate, from_location, to_location) VALUES ('.$box['product_id'].','.$box['size_id'].','.$box['items'].',NOW(),'.$box['location_id'].','.$_POST['location_id'].')');
+            }
         }
 
         // Undelete a box if it is scanned
@@ -74,41 +110,54 @@ if (!$_POST['qr_id']) {
             trigger_error($message);
         }
     }
+    // After a state is changed to scrap or lost, changes are not allowed until the state is changed again
+    if (!$is_lost && !$is_scrap) {
+        // keys of POST to be saved
+        $savekeys = ['box_id', 'product_id', 'size_id', 'items', 'location_id', 'comments', 'qr_id'];
 
-    // keys of POST to be saved
-    $savekeys = ['box_id', 'product_id', 'size_id', 'items', 'location_id', 'comments', 'qr_id'];
-    if (!$new) {
-        $savekeys[] = 'id';
-    }
-
-    $handler = new formHandler('stock');
-
-    $id = $handler->savePost($savekeys);
-    // related trello https://trello.com/c/XjNwO3sL
-    // add remove tagsand insert the new tags when creating the new box
-    db_query('DELETE FROM tags_relations WHERE object_id = :stock_id AND object_type = "Stock"', [':stock_id' => $id]);
-
-    if ($_POST['tags']) {
-        $query = 'INSERT IGNORE INTO tags_relations (tag_id, object_type, `object_id`) VALUES ';
-
-        $params = [];
-        $tags = $_POST['tags'];
-        for ($i = 0; $i < sizeof($tags); ++$i) {
-            $query .= "(:tag_id{$i}, 'Stock', :stock_id)";
-            $params = array_merge($params, ['tag_id'.$i => $tags[$i]]);
-            if ($i !== sizeof($tags) - 1) {
-                $query .= ',';
-            }
+        if (!$new) {
+            $savekeys[] = 'id';
         }
 
-        $params = array_merge($params, ['stock_id' => $id]);
-        db_query($query, $params);
+        $handler = new formHandler('stock');
+
+        $id = $handler->savePost($savekeys);
+        // related trello https://trello.com/c/XjNwO3sL
+        // add remove tagsand insert the new tags when creating the new box
+        db_query('DELETE FROM tags_relations WHERE object_id = :stock_id AND object_type = "Stock"', [':stock_id' => $id]);
+
+        if ($_POST['tags']) {
+            $query = 'INSERT IGNORE INTO tags_relations (tag_id, object_type, `object_id`) VALUES ';
+
+            $params = [];
+            $tags = $_POST['tags'];
+            for ($i = 0; $i < sizeof($tags); ++$i) {
+                $query .= "(:tag_id{$i}, 'Stock', :stock_id)";
+                $params = array_merge($params, ['tag_id'.$i => $tags[$i]]);
+                if ($i !== sizeof($tags) - 1) {
+                    $query .= ',';
+                }
+            }
+
+            $params = array_merge($params, ['stock_id' => $id]);
+            db_query($query, $params);
+        }
     }
 
     // Log qr box connection in history table
     if ($new) {
         simpleSaveChangeHistory('qr', $_POST['qr_id'], 'QR code associated to box.', [], ['int' => $id]);
     }
+
+    // Update the box state if the state changes
+    if (!$new && $newboxstate['box_state_id'] != $box['box_state_id']) {
+        db_query('UPDATE stock SET box_state_id = :box_state_id, ordered = NULL, ordered_by = NULL, picked = NULL, picked_by = NULL, modified = NOW() WHERE id = :id', ['box_state_id' => $newboxstate['box_state_id'],  'id' => $id]);
+        simpleSaveChangeHistory('stock', $box['id'], 'changed box state from '.$box['box_state_name'].' to '.$newboxstate['box_state_name']);
+    } elseif ($new && 'Instock' !== $newboxstate['box_state_name']) {
+        db_query('UPDATE stock SET box_state_id = :box_state_id, ordered = NULL, ordered_by = NULL, picked = NULL, picked_by = NULL, modified = NOW(), modified_by = :user_id WHERE id = :id', ['box_state_id' => $newboxstate['box_state_id'],  'id' => $id, 'user_id' => $_SESSION['user']['id']]);
+        simpleSaveChangeHistory('stock', $box['id'], 'changed box state to '.$newboxstate['box_state_name']);
+    }
+
     $box = db_row('SELECT s.*, CONCAT(p.name," ",g.label) AS product, l.label AS location FROM stock AS s LEFT OUTER JOIN products AS p ON p.id = s.product_id LEFT OUTER JOIN genders AS g ON g.id = p.gender_id LEFT OUTER JOIN locations AS l ON l.id = s.location_id WHERE s.id = :id', ['id' => $id]);
 
     return [$new, $box, $message];
